refactored components to respective folders and fixed overlay issue with scrolling on posts after activating search menu
