--- src/components/Toolbar/Toolbar.css
@@ -1,209 +1,0 @@
-
-
-.nav-container {
-    height: 44x;
-    z-index: 300;
-    position: fixed;
-    left: 0;
-    top: 0;
-    right: 0;
-}
-
-
-.link-search {
-    background: url('../images/search-icon.svg');
-    display: block;
-    background-position: center;
-    background-size: 18px;
-    height: 44px;
-    width: 20px;
-    background-repeat: no-repeat;
-}
-
-
-
-
-
-
-/* Hidden Items */
-
-.mobile-nav,
-.mobile-search-container {
-    display: none;
-}
-
-
-/* Media Queries */
-
-@media (max-width: 1100px) {
-    .nav-container nav {
-        width: 100%;
-    }
-}
-
-@media (max-width: 768px) {
-    .nav-container .desktop-nav {
-        flex-direction: column;
-        position: fixed;
-        top: 0;
-        left: 0;
-        right:0;
-        width: 100%;
-        height: 0vh;
-        background: #000;
-        justify-content: start;
-        overflow: hidden;
-        z-index: -1;
-        transition: all 1000ms ease;
-    }
-
-    .nav-container.active .desktop-nav {
-        height: 100vh;
-    }
-
-    /* Mobile Nav */
-
-    nav .mobile-nav {
-        display: flex;
-        width: 100%;
-        justify-content: space-between;
-        list-style: none;
-    }
-
-    /* Search Box */
-
-    .mobile-search-container input {
-        width: 100%;
-        padding: 4px 36px;
-        font-family: "Inter", sans-serif;
-        font-size: 17px;
-        background: #1b1b1b;
-        border: 0;
-        color: #fff;
-        border-radius: 8px;
-        outline: none;
-    }
-
-    .mobile-search-container {
-        position: relative;
-        padding: 16px 16px 16px 16px;
-        border-bottom: 1px solid #616161;
-        display: flex;
-        align-items: center;
-        opacity: 0;
-        transition: all 600ms ease;
-    }
-
-    .nav-container.active .mobile-search-container {
-        transform: rotateX(0deg);
-        margin-top: 6px;
-        opacity: 1;
-    }
-
-    .mobile-search-container .link-search {
-        position: absolute;
-        left: 24px;
-        opacity: .5;
-        background-size: 15px;
-    }
-
-    /* Nav Move Up */
-
-    .nav-container nav.move-up {
-        margin-top: -70px;
-    }
-
-    /* Cancel Button */
-
-    .mobile-search-container .cancel-btn {
-        color: #ffa929;
-        font-size: 17px;
-        font-weight: 400;
-        cursor: pointer;
-        text-align: end;
-        width: 0px;
-        overflow: hidden;
-        transition: all 400ms ease;
-    }
-
-    .mobile-search-container .search-bar.active + .cancel-btn {
-        width: 74px;
-    }
-
-    .mobile-search-container .search-bar {
-        flex: 1;
-    }
-
-    /* Desktop Nav Move Down */
-
-    nav .desktop-nav.move-down li:first-child {
-        margin-top: 150px;
-    }
-
-    nav .desktop-nav.move-down li {
-        opacity: 0;
-        pointer-events: none;
-    }
-
-    
-    /* Quick Links */
-
-    .mobile-search-container .search-bar.active ~ .search-results {
-        top: 80px;
-        opacity: 1;
-        pointer-events: auto;
-        transition-delay: 40ms;
-    }
-
-    .mobile-search-container .search-results {
-        position: absolute;
-        left: 0;
-        right: 0;
-        padding: 0 32px;
-        opacity: 0;
-        pointer-events: none;
-        top: 10px;
-        transition: all 400ms ease;
-    }
-
-    .mobile-search-container .search-results ul {
-        list-style: none;
-    }
-
-    .mobile-search-container .search-results h2 {
-        color: #86868b;
-        font-size: 12px;
-        text-transform: uppercase;
-        font-weight: 400;
-    }
-
-    .mobile-search-container .search-results ul li a {
-        padding: 16px 0;
-        display: inline-block;
-        text-decoration: none;
-        color: #fff;
-        font-size: 14px;
-        font-weight: 400;
-        border-bottom: 1px solid #616161;
-        width: 100%;
-    }
-
-    .mobile-search-container .search-results ul li a:hover {
-        color: #2997ff;
-    }
-
-    /* Nav Animation */
-
-    .nav-container nav {
-        transition: all 400ms ease;
-    }
-}
-
-
-input,
-div,
-span,
-a {
-    -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
-}
-
